\name{ss}
\alias{ss}
\alias{ad}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Sum of Squared deviations from the mean and sum of Absolute Deviations from the median}
\description{Functions to compute Sum of Squared deviations from the mean and sum of Absolute Deviations from the median.}
\usage{
ss(x)
ad(x)
}
\arguments{
  \item{x}{A numeric vector.}
}
\value{
Sum of Squared deviations from the mean or sum of Absolute Deviations from the median.}
\author{\enc{Aleš Žiberna}{Ales Ziberna}}
\keyword{univar}% at least one, from doc/KEYWORDS
